adding start to moving genes into own model
