+++ yeastphenome/apps/genes/templates/genes/detail.html
@@ -1,0 +1,115 @@
+{% extends "base/wide.html" %}
+{% load static %}
+{% load my_filters %}
+{% block title %}Genes / {{ gene }} / {{ SITE_NAME }}{% endblock %}
+{% block content %}
+    <style>
+        #dataset-genes {
+            font: 10px sans-serif;
+        }
+
+        .bar rect {
+            shape-rendering: crispEdges;
+            cursor: pointer;
+        }
+
+        .bar text {
+            fill: #999999;
+        }
+
+        .axis path, .axis line {
+            fill: none;
+            stroke: #000;
+            shape-rendering: crispEdges;
+        }
+
+        #scrollable-dropdown-menu .tt-menu {
+            max-height: 250px;
+            overflow-y: auto;
+        }
+    </style>
+    <div class="row">
+        <div class="col-md-12">
+            <h1>{{ gene }}</h1>
+            <table class="table">
+                <tr>
+                    <th width="20%" scope="row">Standard Name:</th>
+                    <td>{{ gene.common_name }}</td>
+                </tr>
+                <tr>
+                    <th width="20%" scope="row">Systematic name:</th>
+                    <td>{{ gene.systematic_name }}</td>
+                </tr>
+                <tr>
+                    <th width="20%" scope="row">SGD ID:</th>
+                    <td><a target="_blank" href="https://yeastgenome.org/locus/{{ gene.primary_sgdid }}"
+                           class="external">{{ gene.primary_sgdid }}</a></td>
+                </tr>
+                <tr>
+                    <th width="20%" scope="row">Aliases:</th>
+                    <td>{% for alias in gene.aliases.all %}{{ alias.name }}{% if not forloop.last %}, {% endif %}{% endfor %}</td>
+                </tr>
+                <tr>
+                    <th width="20%" scope="row">Description</th>
+                    <td>{{ gene.description }}</td>
+                </tr>
+                <tr>
+                    <th width="20%" scope="row">Name Description</th>
+                    <td>{{ gene.common_name_explanation }}</td>
+                </tr>
+            </table>
+        </div>
+    </div>
+
+    {% if datasets_top or datasets_bottom %}
+        <div class="row" style="padding-top:20px" id='datasets'>
+            <div class="col-md-12">
+                <h2 style="padding-bottom:20px">Phenotypic Scores</h2>
+                <ul class="nav nav-tabs">
+                    {% if datasets_top %}
+                        <li class="active"><a data-toggle="tab" href="#datasets_top">Top 10</a></li>{% endif %}
+                    {% if datasets_bottom %}
+                        <li><a data-toggle="tab" href="#datasets_bottom">Bottom 10</a></li>{% endif %}
+                </ul>
+
+                <div class="tab-content">
+                    {% if datasets_top %}
+                        <div id="datasets_top" class="tab-pane fade in active" style="padding-top:20px">
+                            {% include "datasets/data_table.html" with datasets=datasets_top table_id="datasets_top_table" %}
+                        </div>{% endif %}
+                    {% if datasets_bottom %}
+                        <div id="datasets_bottom" class="tab-pane fade" style="padding-top:20px">
+                            {% include "datasets/data_table.html" with datasets=datasets_bottom table_id="datasets_bottom_table" %}
+                        </div>{% endif %}
+                </div>
+            </div>
+        </div>{% endif %}
+
+    {% if sims_top or sims_bottom %}
+        <div class="row" style="padding-top:20px" id='similar'>
+            <div class="col-md-12">
+                <h2 style="padding-bottom:20px">Phenotypic Similarities</h2>
+                <ul class="nav nav-tabs">
+                    {% if sims_top %}
+                        <li class="active"><a data-toggle="tab" href="#top10">Top 10</a></li>{% endif %}
+                    {% if sims_bottom %}
+                        <li><a data-toggle="tab" href="#bottom10">Bottom 10</a></li>{% endif %}
+                </ul>
+
+                <div class="tab-content">
+                    {% if sims_top %}
+                        <div id="top10" class="tab-pane fade in active" style="padding-top:20px">
+                            {% include "genes/similar_table.html" with scores=sims_top table_id="similarTable" %}
+                        </div>
+                    {% endif %}
+                    {% if sims_bottom %}
+                        <div id="bottom10" class="tab-pane fade" style="padding-top:20px">
+                            {% include "genes/similar_table.html" with scores=sims_bottom table_id="nonsimilarTable" %}
+                        </div>
+                    {% endif %}
+                </div>
+            </div>
+        </div>{% endif %}
+
+{% endblock %}
+
