+++ yeastphenome/apps/genes/templates/genes/genes_table.html
@@ -1,0 +1,9 @@
+<table id="papersTable" class="table table-condensed table-hover" width="100%">
+        <thead>
+        <tr>
+            <th class="text-nowrap" width="30%">Gene</th>
+            <th width="30%">Aliases</th>
+            <th width="30%">SGDID</th>
+        </tr>
+        </thead>
+</table>
