+++ yeastphenome/apps/genes/templates/genes/similar_table.html
@@ -1,0 +1,19 @@
+<table id="{{ table_id }}" class="table table-condensed table-hover" width="100%">
+    <thead>
+    <tr>
+        <th class="text-nowrap" width="30%">Gene</th>
+        <th width="30%">Correlation Mean</th>
+        <th width="30%">Correlation Std. Dev.</th>
+    </tr>
+    </thead>
+    <tbody>
+    {% for sim in scores %}
+        <tr>
+            <td><a href="{% url 'genes:detail' sim.gene2.id %}">{{ sim.gene2 }}</a></td>
+            <td>{{ sim.score|stringformat:".2f" }}</td>
+            <td>{{ sim.pvalue|stringformat:".2f" }}</td>
+        </tr>
+    {% endfor %}
+</table>
+<a class="btn btn-info" style="margin-right:5px" role="button" href="{% url 'genes:download_gene_similarities' gene.id %}">Download all</a>
+<a class="btn btn-success" role="button" href="{% url 'genes:similar_genes' gene.id %}">View all</a>
